Move autosetup out of init.py
