+++ homeassistant/components/ihc/const.py
@@ -1,8 +1,11 @@
 """IHC component constants."""
+from homeassistant.const import Platform
 
 ATTR_IHC_ID = "ihc_id"
 ATTR_VALUE = "value"
 ATTR_CONTROLLER_ID = "controller_id"
+
+AUTO_SETUP_YAML = "ihc_auto_setup.yaml"
 
 CONF_AUTOSETUP = "auto_setup"
 CONF_BINARY_SENSOR = "binary_sensor"
@@ -22,6 +25,12 @@ CONF_XPATH = "xpath"
 DOMAIN = "ihc"
 
 IHC_CONTROLLER = "controller"
+IHC_PLATFORMS = (
+    Platform.BINARY_SENSOR,
+    Platform.LIGHT,
+    Platform.SENSOR,
+    Platform.SWITCH,
+)
 
 SERVICE_SET_RUNTIME_VALUE_BOOL = "set_runtime_value_bool"
 SERVICE_SET_RUNTIME_VALUE_FLOAT = "set_runtime_value_float"
